requirements-mcp.txt
# MCP dependencies
mcp>=1.0.0

# Core Google API dependencies
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0

# AI dependencies
openai>=1.0.0

# Document processing
python-docx>=0.8.0
PyPDF2>=3.0.0
openpyxl>=3.1.0

# Async and scheduling
aiofiles>=23.0.0
schedule>=1.2.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dateutil>=2.8.0